moved datetime/dayjs related functions to own file
