+++ src/components/Plugin/CommentBox/ReplyBlock.vue
@@ -6,7 +6,7 @@ import { useWeb3 } from '@/composables/useWeb3';
 import { signMessage } from '@snapshot-labs/snapshot.js/src/utils/web3';
 import { getInstance } from '@snapshot-labs/lock/plugins/vue3';
 import { useI18n } from 'vue-i18n';
-import { relativeTimeFromTimestamp } from '@/helpers/utils';
+import { relativeTimeFromTimestamp } from '@/helpers/datetime';
 const { t } = useI18n();
 const auth = getInstance();
 const { modalOpen, modalAccountOpen } = useModal();
